refactor(input): Move input-info
Move input info into its own folder, since it is no longer exclusive to ux-input.
