+++ src/input-info/ux-input-info-theme.css
